requirements.txt
flask 
pickle
scikit-learn 
nltk 
python-dotenv 
openai